Carl...a quick pass through Volume 7.  You probably already have this on your action list, but we need to standardize the new V1.2 Volume list to include Volumes 13 (GPKG EXT?) and Volume 14 (renamed Multi-Spectral extension?)
For now, I suggest we leave the UHRB section alone.  The graphics in 
there illustrate the points.  I don't think we will have resolution on 
re-establishing public links to the UHRB Object Model and UHRB On Disk 
Format documents for awhile, but I also don't think anyone has, or plans 
to, implement the UHRB building interior codes and sub-codes.